changes from comments and directory restructure
