+++ platform-operator/constants/constants.go
@@ -14,9 +14,6 @@ const VerrazzanoInstallNamespace = "verrazzano-install"
 // VerrazzanoMonitoringNamespace is the namespace for monitoring components
 const VerrazzanoMonitoringNamespace = "verrazzano-monitoring"
 
-// VerrazzanoMonitoringStackNamespace is the namespace for the monitoring stack
-const VerrazzanoMonitoringStackNamespace = "verrazzano-monitoring-stack"
-
 // Verrazzano is the name of the Verrazzano secret in the Verrazzano system namespace
 const Verrazzano = "verrazzano"
 
@@ -165,11 +162,9 @@ const UpgradeRetryVersion = "verrazzano.io/upgrade-retry-version"
 const ObservedUpgradeRetryVersion = "verrazzano.io/observed-upgrade-retry-version"
 
 // VerrazzanoDevComponentAnnotationName is the name of the annotation to be used on a ConfigMap,
-// which indicates that the ConfigMap is a VZ stack configuration. The value of the annotation is
-// the VZ stack name e.g. "monitoring"
-const VerrazzanoDevComponentAnnotationName = "verrazzano.io/devComponent"
-
-const VerrazzanoDevComponentChartAnnotationName = "verrazzano.io/devComponentChart"
+// which indicates that the ConfigMap is used to install a dev component. The value of the annotation is
+// the dev component name unless the name field is otherwise definied in the ConfigMap data
+const VerrazzanoDevComponentAnnotationName = "verrazzano.io/component"
 
 // NGINXControllerServiceName is the nginx ingress controller name
 const NGINXControllerServiceName = "ingress-controller-ingress-nginx-controller"
