+++ platform-operator/controllers/configmaps/overrides/controller_utils_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, 2023,  Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package overrides_configmaps
+package overrides
 
 import (
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
